src/requirements.txt
django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.47
redis==5.1.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
django-filter==24.3
Pillow==10.4.0
gunicorn==23.0.0
pre-commit==3.8.0
django-modeltranslation==0.19.9
